requirements.txt
alembic==1.12.0
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.3
click==8.1.7
flasgger==0.9.7.1
Flask==3.0.0
Flask-Admin==1.6.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
loguru==0.7.2
Mako==1.2.4
MarkupSafe==2.1.3
mistune==3.0.2
packaging==23.2
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==3.0.0
WTForms==3.1.0





